Added engine-based cutscene control
The new class TextBox represents a box of text that can easily be
controlled through the level editor. When triggered, the game pauses
and the line(s) of text stored as properties of the connection are
displayed consecutively on the screen. The user can press enter (or
down) to scroll through the messages, and when they are done, the game
resumes. Sound can also easily be played as part of this, which is nice.